refactor: Moved meta data log directory to src
