+++ lib/src/meta_data_logs/ingredient_parsing_failure_meta_data_log.dart
