+++ lib/src/meta_data_logs/meta_data_log.dart
@@ -2,8 +2,8 @@ library meta_data_log;
 
 import 'package:easy_localization/easy_localization.dart';
 
-import '../../../l10n/locale_keys.g.dart';
-import '../recipe_parsing_job.dart';
+import '../../l10n/locale_keys.g.dart';
+import '../models/recipe_parsing_job.dart';
 
 part 'amount_parsing_failure_meta_data_log.dart';
 part 'complete_failure_meta_data_log.dart';
