+++ lib/src/recipe_scripts/wordpress_ingredient_parsing.dart
@@ -1,8 +1,8 @@
 import 'package:html/dom.dart';
 
+import '../meta_data_logs/meta_data_log.dart';
 import '../models/ingredient.dart';
 import '../models/ingredient_parsing_result.dart';
-import '../models/meta_data_logs/meta_data_log.dart';
 import 'parsing_helper.dart';
 
 /// Parses an html [Element] representing an [Ingredient].
